requirements.txt
streamlit
python-dotenv
google-generativeai
pdfplumber
python-docx
